split dsapp to pages & add llm_logs showing
